+++ test/fixture/define/dynamic-import/input.js
@@ -1,0 +1,7 @@
+define(function () {
+  return function fetch () {
+    return import('./foo')
+        .then(() => { console.log('success') })
+        .catch(() => { console.log('error') })
+  }
+})