+++ test/fixture/define/dynamic-import/output.js
@@ -1,0 +1,5 @@
+export default function fetch () {
+  return import('./foo')
+      .then(() => { console.log('success') })
+      .catch(() => { console.log('error') })
+}