+++ test/spec/convert.js
@@ -313,6 +313,14 @@ test('(convert) it works for multiple returns', assert => {
   assert.truthy(convert('define/multiple-returns'))
 })
 
+test('(convert) it is a noop for files with dynamic imports', assert => {
+  assert.truthy(convert('noop/dynamic-import'))
+})
+
+test.skip('(convert) it does not break on files with dynamic import', assert => {
+  assert.truthy(convert('define/dynamic-import'))
+})
+
 test.skip('it leaves single line comments if required', assert => {
   assert.truthy(convert('define/single-line-comments', { comments: true }))
 })